fixed a lil bug, added a larger example input data
that data is copied from my teacher who gave me this task, though i changed it a lot
(and obv reformatted it to be zon instead of whatever he used).
please dont try to dox me through this 3:
after this school year ill be gone anyway, so there wouldnt be much ull gain.